Add a proper distutils setup, move into module, document install
The module was distributed as a plain Python file that users were asked to copy
into their projects. This should be avoided. Instead it should be installed
using distutils as a namespaced module, so one copy can be updated globally across
the system.

The code is moved into the __init__.py of the shiftplanning module. This is
completely backward compatible for existing code, but allows the module to be
split into separate files later, extra modules added without namespace issues,
etc.

A quick note is added to the readme to document how to install it sanely. In
particular it advises the user to install simplejson using easy_install or pip
if possible, rather than hand-copying it.